requirements.txt
future~=0.18.3
requests~=2.31.0
numpy~=1.26.3
setuptools~=69.0.3
PyQt5~=5.15.10
certifi~=2023.11.17
imageio~=2.33.1
pillow~=10.2.0
moviepy~=1.0.3
modulegraph~=0.19.6
py2app~=0.28.6
macholib~=1.16.3
Eel~=0.16.0
gevent~=23.9.1
bottle~=0.12.25
pyparsing~=3.1.1
whichcraft~=0.6.1
pip~=23.3.2
wheel~=0.41.2
greenlet~=3.0.3
chardet~=5.2.0
proglog~=0.1.10
tqdm~=4.66.1
decorator~=4.4.2
pyarmor~=8.4.6
altgraph~=0.17.4
icnsutil~=1.1.0
rubicon-objc
idna~=3.6
urllib3~=2.1.0
packaging~=23.2
pyqt5
moviepy
pillow
py2app
pyautogui